moved input utils to object
